refactor(template): move templateRequiresConfirmation from router to template module
